requirements.txt
aniso8601==9.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
Flask==2.0.3
flask-restx==0.5.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.0.0
MarkupSafe==2.0.1
pkg-resources==0.0.0
pyrsistent==0.18.0
python-decouple==3.6
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
zipp==3.6.0
